moved updated AHN_FlowTools file into core folder
